requirements.txt
streamlit==1.28.1
langchain==0.0.335
langchain-community==0.0.6
sentence-transformers==2.2.2
python-docx==0.8.11
pyyaml==6.0.1
transformers==4.35.2
torch==2.1.1
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0

# Weaviate
weaviate-client==3.24.2

# API clients
groq==0.4.1
google-generativeai==0.3.2

# Document processing
python-docx==0.8.11
docx2txt==0.8

# Additional ML libraries
sentence-transformers==2.2.2
accelerate==0.24.1

# FastAPI Backend
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# Additional utilities
aiofiles==23.2.1
requests==2.31.0
psutil==5.9.6

# Testing
pytest==7.4.3
httpx==0.25.2